made entry grabber its own function
