+++ src/scripts/entryGrabber.js
@@ -1,18 +1,19 @@
-// import API from "./data.js"
 
-// // grabs new entry data 
-
+// grabs new entry data 
 
 
-// document.querySelector("#myJournalButton").addEventListener("click", event => {
 
-//     const journalEntryObj = {}
+const entryGrabber = () => {
 
-//     journalEntryObj.date = document.querySelector("#journalDate").value
-//     journalEntryObj.concepts = document.querySelector("#journalConcepts").value
-//     journalEntryObj.entry = document.querySelector("#journalEntry").value
-//     journalEntryObj.mood = document.querySelector("#mood").value
+    const journalEntryObj = {}
 
-//     console.log(journalEntryObj)
-//     API.saveJournalEntry(journalEntryObj)
-// })
+    journalEntryObj.date = document.querySelector("#journalDate").value
+    journalEntryObj.concepts = document.querySelector("#journalConcepts").value
+    journalEntryObj.entry = document.querySelector("#journalEntry").value
+    journalEntryObj.mood = document.querySelector("#mood").value
+
+   return journalEntryObj
+}
+
+
+export default entryGrabber
